refactor(ui): organise useInputFieldTemplate hooks again & add useInputFieldTemplateSafe
