+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/NodeFieldElementViewMode.tsx
@@ -5,7 +5,7 @@ import { InputFieldRenderer } from 'features/nodes/components/flow/nodes/Invocat
 import { useContainerContext } from 'features/nodes/components/sidePanel/builder/contexts';
 import { NodeFieldElementLabel } from 'features/nodes/components/sidePanel/builder/NodeFieldElementLabel';
 import { useInputFieldDescriptionSafe } from 'features/nodes/hooks/useInputFieldDescriptionSafe';
-import { useInputFieldTemplateOrThrow } from 'features/nodes/hooks/useInputFieldTemplateOrThrow';
+import { useInputFieldTemplateOrThrow, useInputFieldTemplateSafe } from 'features/nodes/hooks/useInputFieldTemplate';
 import type { NodeFieldElement } from 'features/nodes/types/workflow';
 import { NODE_FIELD_CLASS_NAME } from 'features/nodes/types/workflow';
 import Linkify from 'linkify-react';
@@ -29,12 +29,12 @@ export const NodeFieldElementViewMode = memo(({ el }: { el: NodeFieldElement }) 
   const { id, data } = el;
   const { fieldIdentifier, showDescription } = data;
   const description = useInputFieldDescriptionSafe(fieldIdentifier.nodeId, fieldIdentifier.fieldName);
-  const fieldTemplate = useInputFieldTemplateOrThrow(fieldIdentifier.nodeId, fieldIdentifier.fieldName);
+  const fieldTemplate = useInputFieldTemplateSafe(fieldIdentifier.nodeId, fieldIdentifier.fieldName);
   const containerCtx = useContainerContext();
 
   const _description = useMemo(
-    () => description || fieldTemplate.description,
-    [description, fieldTemplate.description]
+    () => description || fieldTemplate?.description || '',
+    [description, fieldTemplate?.description]
   );
 
   return (
@@ -45,22 +45,39 @@ export const NodeFieldElementViewMode = memo(({ el }: { el: NodeFieldElement }) 
       data-parent-layout={containerCtx.layout}
       data-with-description={showDescription && !!_description}
     >
-      <FormControl flex="1 1 0" orientation="vertical">
-        <NodeFieldElementLabel el={el} />
-        <Flex w="full" gap={4}>
-          <InputFieldRenderer
-            nodeId={fieldIdentifier.nodeId}
-            fieldName={fieldIdentifier.fieldName}
-            settings={data.settings}
-          />
-        </Flex>
-        {showDescription && _description && (
-          <FormHelperText sx={linkifySx}>
-            <Linkify options={linkifyOptions}>{_description}</Linkify>
-          </FormHelperText>
-        )}
-      </FormControl>
+      <NodeFieldElementViewModeContent el={el} />
     </Flex>
   );
 });
 NodeFieldElementViewMode.displayName = 'NodeFieldElementViewMode';
+
+const NodeFieldElementViewModeContent = memo(({ el }: { el: NodeFieldElement }) => {
+  const { data } = el;
+  const { fieldIdentifier, showDescription } = data;
+  const description = useInputFieldDescriptionSafe(fieldIdentifier.nodeId, fieldIdentifier.fieldName);
+  const fieldTemplate = useInputFieldTemplateOrThrow(fieldIdentifier.nodeId, fieldIdentifier.fieldName);
+
+  const _description = useMemo(
+    () => description || fieldTemplate.description,
+    [description, fieldTemplate.description]
+  );
+
+  return (
+    <FormControl flex="1 1 0" orientation="vertical">
+      <NodeFieldElementLabel el={el} />
+      <Flex w="full" gap={4}>
+        <InputFieldRenderer
+          nodeId={fieldIdentifier.nodeId}
+          fieldName={fieldIdentifier.fieldName}
+          settings={data.settings}
+        />
+      </Flex>
+      {showDescription && _description && (
+        <FormHelperText sx={linkifySx}>
+          <Linkify options={linkifyOptions}>{_description}</Linkify>
+        </FormHelperText>
+      )}
+    </FormControl>
+  );
+});
+NodeFieldElementViewModeContent.displayName = 'NodeFieldElementViewModeContent';
